no-code/google-functions/transcript/requirements.txt
google-cloud-storage
google-cloud-speech
ffmpeg-python